New JSP Rule: DuplicateJspImport
Checks for Duplicate imports in JSP's.

Found that, though CompilationUnit is inside the AST, it's not being visited, so created a short term hack to override a parent method so we know when the head of each file parse is - that way we can purge the list of imports.

Once CompilationUnits are visited properly I'll fix this


git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk/pmd@4353 51baf565-9d33-0410-a72c-fc3788e3496d