Checkin of the Type Resolution Facility
Re-implemented LooseCoupling as a new rule - new or changed type resolution rules should go in typeresolution to be separated.

I have not added this to 39.xml!

Sorry about the lack of comments, will go through and try and get important things commented.

There's a couple TODO's that if addressed should help performance.

Remember, to run compiled classes must be in the CLASSPATH!


git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk/pmd@4744 51baf565-9d33-0410-a72c-fc3788e3496d