Fixed this JUnit test
Since it's a mocked class being reported on, it's reporting n/a for the file name. There's code in the PapariTextRenderer class that parses out the class name - based on substring of the File Separator parameter - On *Nix that's the / character.

The test passes for me on both linux and Windows now


git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk/pmd@5025 51baf565-9d33-0410-a72c-fc3788e3496d