Applied patch from : 1751053 - New XSL a la Checkstyle/Maven.
Added an XSLTRenderer which inherits from XMLRenderer. 
Once the XMLRenderer has done his job, the XML result will be transformed using the patch's XSTL.
It possible to use this renderer with other stylesheet, but not from the command line nor ant.

This new formatter does NOT override the old HTML renderer, to use this new one, we have to specify, 
in command line, the "betterhtml" renderer:

To quicly test it on PMD itself:
$java net.sourceforge.pmd.PMD src betterhtml rulesets/unusedcode.xml          
 
As i had to write a new renderer, i don't feel we should remove the old one for the moment.

(as always, i forgot the changelog)
Also, note that this note work with Ant, i'm working on it...


git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk/pmd@5490 51baf565-9d33-0410-a72c-fc3788e3496d