Added a 'since' attribute to Rules.  This is the version of PMD in which a Rule was first delivered.  This can help IDE users know what rules are new to PMD (e.g. I just upgraded my IDE plugin which was using PMD a.b and and PMD is now x.y, what rules are newer than a.b?).
All built-in PMD rules should have a value for this (added a JUnit test to that effect).  We need to go back and annotate all PMD rules with their original release version.  Looks like we have easy records going back to PMD 3.3, prior to that will require a bit of research in the SVN history I imagine.

git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk/pmd@5785 51baf565-9d33-0410-a72c-fc3788e3496d