Allow the ClassLoader used by PMD for Type Resolution to be specified.  This ClassLoader can be distinctly different than the ClassLoader used to load PMD itself.
IDE plugins can create ClassLoaders representing the "project" settings in the IDE, providing PMD with that ClassLoader, such that PMD can truly see the classes being represented in the IDE.  In the specific case of Eclipse, the IDE plugin runs in a ClassLoader which does not contain the details of the "project" settings, which means without this feature PMD would never be able to do proper Type Resolution.

I've not had a chance to benchmark this change.  I expect there to be some performance changes given that 'static' reference to the PMDASMClassLoader is no longer used.  Further tuning may need to be done.

git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk/pmd@5819 51baf565-9d33-0410-a72c-fc3788e3496d