optimization: DoNotCallGarbageCollectionExplicitly rewritten. The new implementation only takes 40% of the time.
Note: when writing a new rule, using command line option "-benchmark" on a few rules
can give an indication on how the rule compares to others. To get the full picture,
use the rulesets/internal/all-java.xml ruleset with "-benchmark".

This rule was the slowest when running on the jdk source tree: it took 69s to complete.
The new version only takes 27s.


git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk/pmd@5851 51baf565-9d33-0410-a72c-fc3788e3496d