Eclipse formatter settings, based on Sun Java conventions, with minor changes to line length and no formatting of comments.  See comment at top of the file.
The annoying part of this is the treatment of indentation.  Sun appears to like to mix tabs and spaces.  I prefer to go with one or the other, but I don't really care as long as all the code in PMD is the same so I _can_ use my IDEs code formatting capabilities.

Feel free to suggest alternate settings, or even alternate tools.  As long as we can come up with a uniform standard that all devs can apply in their development environments.

git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk/pmd@5938 51baf565-9d33-0410-a72c-fc3788e3496d