Rule refactoring into proper Language packages.
Also renamed Rules to following standard convention and organization, and enhanced the RuleSetFactoryTest case to assert this convention.  For example, if the non-XPath Rule is XxxYyy/zzz.xml, then the rule class should be:

   lang.${lang}.rule.${zzz}.${XxxYyy}Rule

Ahh... wonderous consistency. :)

We're getting closer, much of the big stuff is now done, but still a fair bit to do, see changelog for details.

git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk/pmd@5958 51baf565-9d33-0410-a72c-fc3788e3496d