Additional changes to use common JavaCC infrastructure across the Languages.  JSP stuff no longer extends from the Java stuff.  Also, the files TokenMgrError and ParseException are regenerated from scratch now.  Only the Token class is not regenerated, due to customizations.  We're going to have to keep an eye on JavaCC changes which might affect this class so we can regenerate and apply customizations again, as needed.
Added the Ant build target 'dummyjjtree' to regenerate the common JavaCC components into the lang.ast package.  Also, added a target 'alljjtree' to generate all the JavaCC related code, so that upgrades to JavaCC can applied to all PMD code, and not just a single language at a time.

FYI, JavaCC seems to be putting checksums in files, and these are calculated _with_ the EOL characters, so you can get spurious checksum differences between files when the prior version was generated on a platform with a different EOL (e.g. Windows versus *nix).  Verify the "changed" files are really changed before committing.

git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk/pmd@5968 51baf565-9d33-0410-a72c-fc3788e3496d