Enhance typeof function to support not passing an Attribute containing the String form of the type.  As of PMD 4.2, not every TypeNode is a direct container for such a Type name.
It is still the case that passing in the @Image attribute will give better results in the absence of a null TypeNode.getType() return result.  Is there a way to do this automagically when appropriate based upon the TypeNode?  We pass the Class up the AST, can we pass the String too?

git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk/pmd@6167 51baf565-9d33-0410-a72c-fc3788e3496d