Modified Parser interface to indicate if a Language can actually be parsed or tokenized only (e.g. C++).  Given that, and the existing of a LanguageVersionHandler on a given Language can be used to drive the availability of a LanguageVersion in the Designer interface.
For the XML Language, implemented the Node.jjtGetParent interface, the RootNode is the DOM node with a Document parent.  Also, fixed issues with the AST wrapping of the XML DOM nodes.  Need to cache the already wrapped nodes so the AST and DOM hierarchies are in 1-1 (isomorphic) relationship, in particular the parent and child relationships still make sense at the Java Object level.  Finally, added a very primitive dumper for XML Language, so the AST nodes can be printed in the Designer.  Feel free to enhance the dumper!

git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk/pmd@6175 51baf565-9d33-0410-a72c-fc3788e3496d