Designer fixes:
LanguageVersion default fixed, was incorrectly using enum ordinal() as index, needs to use index in List of LanguageVersions.

Change to use Jaxen Navigator instance as defined by the LanguageVersion.  For the XML Language, instead of the primitive AST Navigator, use a full blown DOM navigator!  This is how the XPathRule works.


git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk/pmd@6179 51baf565-9d33-0410-a72c-fc3788e3496d