Changes to Rule interface to support a Language value.  Every Rule must have a Language.  All the Java AST Rule base classes will default the Language, and enforce that it cannot be changed to something else (e.g. ImmutableLanguage interface).  XPath Rules must specify a Language in the RuleSet XML.  All existing Rules have had their languages set appropriately.
Also added a minimum and maximum Language Version to Rules.  By default these are null, indicating a Rule applies to all versions of a Language.  Some Rules can be defined to work only on certain Language Version ranges.  It is envisioned this will be used for Java migration related rules, and for future Rules concerning newer Java language feature usage (e.g. Annotations, etc).

Next step is to rework RuleSet to no longer use Language directly, but to defer to it's Rules.


git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk/pmd@6195 51baf565-9d33-0410-a72c-fc3788e3496d