Change the "excludeMarker" to "suppressMarker" to make this consistent with the use of "suppress" elsewhere in PMD, and with the standard Java @SuppressWarnings() annotation.
Remove some of the processFiles() APIs on the PMD class, we really don't need that many of them.  In this regard, the one API that allowed a LanguageVersion to be passed explicitly has been removed, and the RuleContext class update to indicate this is the preferred method of indicating the LanguageVersion to use when processing a source file.

Initial cut at adding a Configuration class, for storing all PMD configuration related items.  So far, I've just moved the stuff that was stored on the PMD class itself to the Configuration class.  Additional work will be done to move the stuff currently in CommandLineOptions and in the Ant task onto the Configuration class.  Then the processFiles methods will be modified to take a Configuration option.  This should prevent the proliferation of processFile() methods in the future, and make it possible to unify some common aspects of the command line/Ant usages.

git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk/pmd@6208 51baf565-9d33-0410-a72c-fc3788e3496d