Changed the dogfood to be all encompassing.  The goal is for every Rule which can be applied to PMD, to be applied automatically.  If a Rule is not to be applied, then it must be explicitly exclude from the dogfood RuleSet.
For now, I've made arbitrary decision about which Rules we will exclude (and thus implicitly which will be included).  For Rules which are to be included, but for which the PMD code is not yet passing the Rule, there is a temporary exclusion (see the numerous TODO comments and the excludes below each such comment).

Finally, the Ant pmdOnPmd target is changed to fail on rule violations (which will break the CC build!).  Essentially, dogfood is not optional anymore, so no more excuses for being sloppy (I'm guilty of this!).  All PMD code must pass our own dogfood.  Plan is to work during the 5.0 timeline to get the code base to be 100% compliant with the "ideal" dogfood RuleSet.

I imagine I won't get disagreement on this change, but perhaps debate/feedback on individual Rules?

git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk/pmd@6254 51baf565-9d33-0410-a72c-fc3788e3496d