Dogfood now contains LooseCoupling.
Refactored usages of LinkedList to use ArrayList, primarily in the DFA area, but a few other places too.  AFAICT, a LinkedList wasn't strictly needed here (not insert/delete performance needs), and if used would likely be a tad bit slower due to the extra object creations which goes on behind the scenes for each item in the List compared to an ArrayList.


git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk/pmd@6321 51baf565-9d33-0410-a72c-fc3788e3496d