Change version property to have choices and labels be identical.
EnumeratedProperty is really odd in that it wants to serialize out the label.  I had assumed it would serialize out the value, and the label was merely used to provide a description in some GUI, and was not the externalized form.  Regardless, a verbose value is not what I want people to have to enter when manually providing a the version for XPathRule in XML.

git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk/pmd@6542 51baf565-9d33-0410-a72c-fc3788e3496d