Remove Language from RuleSet class.  Previous changes which associated Language with Rule are now the driver for processing instead of the RuleSet.
Work still needs to be done to clean up the XML Schema and DTD.  Nothing too complex there.

Slightly more complicated work needs to be done allow PMD to appropriately associate an source file based LanguageVersion on the RuleContext.  Once that is done, it will be technically possible to use a single RuleSet containing Rules for different Languages and actually run it against a mixed set of source files.

git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk/pmd@6546 51baf565-9d33-0410-a72c-fc3788e3496d