Create the RuleSetReferenceId class.  The goal is to provide a single means of identifying a RuleSet for all mechanisms which use this (RuleSetFactory, RuleSet XML, and command line).  This class merges together the functionality of ExternalRuleID and SimpleRuleSetNameMapper.
My original impetus for this change is the following additional feature... support for creating a reference to a Rule in the "current" RuleSet.  This has no external use, but can/will be used to enhance the RuleSetFactory to allow references to a Rule in the current RuleSet, without having to make a fully qualified reference to the current RuleSet.  It is like being able to call a static method defined in the current Java class, without having to prefix the method with the FQN of the current Java class.

I will now work on refactoring everything to use this class.

When done, this change will also make it possible to list a single Rule anywhere a RuleSet name is to be expected, be it in the RuleSet XML, or the PMD command line (a feature requested in the past).


git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk/pmd@6756 51baf565-9d33-0410-a72c-fc3788e3496d